refactor(analysis): migrate cytotrace module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
