+++ src/pyrovelocity/analysis/__init__.py
@@ -1,6 +1,8 @@
 import pyrovelocity.analysis.analyze
+import pyrovelocity.analysis.cytotrace
 
 
 __all__ = [
     "analyze",
+    "cytotrace",
 ]
