+++ src/pyrovelocity/tasks/preprocess.py
@@ -14,7 +14,7 @@ from anndata._core.anndata import AnnData
 from beartype import beartype
 from scipy.sparse import issparse
 
-from pyrovelocity.cytotrace import cytotrace_sparse
+from pyrovelocity.analysis.cytotrace import cytotrace_sparse
 from pyrovelocity.logging import configure_logging
 from pyrovelocity.tasks.data import load_anndata_from_path
 from pyrovelocity.utils import ensure_numpy_array
